LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/access-control-preflight-request-invalid-status-400-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 400
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/access-control-preflight-request-invalid-status.py?code=400 due to access control checks.

PASS Request with 400 preflight response
